Move `MarkdownText` to `Icinga\Module\Icingadb\Widget`
